feat(web): add project-level schedulers entry to sidebar
Bring the per-project scheduler bindings page out of Project Settings
and into the project navigation so members can reach it directly from
the left panel (and the tabbed project header) at
/:workspaceSlug/projects/:projectId/schedulers.

The settings location (/settings/projects/:id/schedulers) is kept as-is.
Both pages now render a shared SchedulerBindingsPanel that hides
install/edit/uninstall controls for non-admins, matching the backend's
admin-only mutation gate.

Also drops redundant inner workspaceSlug/projectId params in the project
navigation hooks (they shadowed the outer scope and tripped lint-staged).
